Move transport layer dependent code from SessionCommon to transport_layer::StreamTransport
